% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{bohn}
\alias{bohn}
\title{Bohn dune layer with 500 m buffer (EPSG:25832)}
\format{
An `sf` object.
}
\source{
source?
}
\usage{
data(dunes_bohn_500mbuffer_epsg25832)
}
\description{
Coastal dune regions from Udo Bohn with 500 m buffer in Projection EPSG:25832.
}
\keyword{datasets}
