% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{bohn_25832}
\alias{bohn_25832}
\title{Bohn dune layer (unprojected)}
\format{
An `sf` object.
}
\source{
source?
}
\usage{
data(dunes_bohn_500mbuffer_25832)
}
\description{
Coastal dune regions from Udo Bohn in Projection EPSG:25832.
}
\keyword{datasets}
