% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{co}
\alias{co}
\title{EEA coastline regions of Europe (EPSG:25832)}
\format{
An `sf` object.
}
\source{
source?
}
\usage{
data(coastline_regions_epsg25832)
}
\description{
Coastline regions from the European Environment Agency in Projection EPSG:25832.
}
\keyword{datasets}
