% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{coastline_regions_epsg25832}
\alias{coastline_regions_epsg25832}
\alias{co}
\title{Regional seas around Europe (EPSG:25832)}
\format{
An `sf` object with 5 polygons.
}
\source{
European Environment Agency: Regional seas around Europe.
}
\usage{
data(coastline_regions_epsg25832)
}
\description{
The regional seas around Europe (Marine Strategy Framework Directive regions)
in EPSG:25832. The dataset holds the `sf` object `co`; its column `COAST_EEA`
gives the coast code of each sea (for example `MED_COAST`). Used to set
`Coast_EEA` in [resy_harmonize_eunis()].
}
\keyword{datasets}
