% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{dunes_bohn_500mbuffer_epsg25832}
\alias{dunes_bohn_500mbuffer_epsg25832}
\alias{bohn}
\title{Coastal vegetation of the Bohn map with a 500 m buffer (EPSG:25832)}
\format{
An `sf` object with 367 polygons.
}
\source{
Bohn, U. et al.: Map of the Natural Vegetation of Europe.
}
\usage{
data(dunes_bohn_500mbuffer_epsg25832)
}
\description{
Polygons of the coastal vegetation (formation P) of the Map of the Natural
Vegetation of Europe (Bohn et al.), buffered by 500 m, in EPSG:25832. The
dataset holds the `sf` object `bohn`; its column `DUNE` marks dune polygons.
Used to set `Dunes_Bohn` in [resy_harmonize_eunis()].
}
\keyword{datasets}
