% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resy_aggregate_taxa.R
\name{resy_aggregate_taxa}
\alias{resy_aggregate_taxa}
\title{Aggregate taxa to expert-system aggregation level}
\usage{
resy_aggregate_taxa(obs, aggs)
}
\arguments{
\item{obs}{A `data.table` with column `TaxonName`.}

\item{aggs}{Named list of aggregations as returned by [resy_parse_expert()].}
}
\value{
Modified `obs` as `data.table`.
}
\description{
Replaces `obs$TaxonName` by aggregated names if they occur in the aggregation
mapping of the expert system. If you use GermanSl or EuroSL, use `taxval` instead.
}
