% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{europe_resolution_60_epsg25832}
\alias{europe_resolution_60_epsg25832}
\title{European countries in EPSG:25832}
\format{
An `sf` object.
}
\source{
source?
}
\usage{
data(europe_resolution_60_epsg25832)
}
\description{
GIS data of the Ecoregions in Projection EPSG:25832.
}
\keyword{datasets}
