% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resy_eval_plot.R
\name{resy_eval_plot}
\alias{resy_eval_plot}
\title{Evaluate and print details for a single plot}
\usage{
resy_eval_plot(res, p, type)
}
\arguments{
\item{res}{A `resy_result` returned by [resy_classify()].}

\item{p}{Plot identifier (`PlotObservationID`) as character; if numeric, treated as
row index in `header`.}

\item{type}{If vegetation-type short code is given, relevant conditions are shown}
}
\description{
Evaluate and print details for a single plot
}
