% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resy_eval_type.R
\name{resy_eval_type}
\alias{resy_eval_type}
\title{Evaluate and print details for a vegetation type}
\usage{
resy_eval_type(res, t)
}
\arguments{
\item{res}{A `resy_result` returned by [resy_classify()].}

\item{t}{Vegetation-type short code (e.g. "R55") or numeric index.}
}
\description{
Evaluate and print details for a vegetation type
}
